css/css-flexbox/flexbox_mix-writing-flex-alignment.html

<!DOCTYPE html>
<html lang="en">
<head>

  <title>Flexbox | Flex alignment and writing modes</title>
  <meta charset="utf-8">
  <meta name="assert" content="Flex alignment and writing modes">
  <link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#pagination">
  <link rel="author" title="Clayton Salem" href="claytonjsalem@gmail.com">
  <style>
    /* Writing mode utility classes. */
    .horizontal-tb { writing-mode: horizontal-tb; }
    .vertical-rl { writing-mode: vertical-rl; }
    .vertical-lr { writing-mode: vertical-lr; }
    .sideways-rl { writing-mode: sideways-rl; }
    .sideways-lr { writing-mode: sideways-lr; }
    /* Positioning utility classes. */
    .start { justify-content: start; }
    .end { justify-content: end;}
    .space-between { justify-content: space-between; }
    .space-around { justify-content: space-around; }
    .space-evenly { justify-content: space-evenly; }
    .flex-end { align-items: flex-end; }
    .flex-start { align-items: flex-start; }
    .baseline { align-items: baseline; }
    .stretch { align-items: stretch; }
    .center { align-items: center; }
    .row { flex-direction: row; }
    .row-reverse { flex-direction: row-reverse; }
    .column { flex-direction: column; }
    .column-reverse { flex-direction: column-reverse; }
    /* A splash of color. */
    .container {
      background-color: #4fbe99;
      height: 100px;
      width: 200px;
    }
    .inner {
      background-color: #3d5d67;
      color: white;
      height: 100px;
      width: 100px;
    }
  </style>
</head>
<body>
  <div id="wrap">
  <div class="container horizontal-tb column-reverse"><div class="inner">horizontal-tb<br>column-reverse</div></div><div class="container horizontal-tb start"><div class="inner">horizontal-tb<br>start</div></div><div class="container horizontal-tb center"><div class="inner">horizontal-tb<br>center</div></div><div class="container horizontal-tb space-around"><div class="inner">horizontal-tb<br>space-around</div></div><div class="container horizontal-tb space-between"><div class="inner">horizontal-tb<br>space-between</div></div><div class="container horizontal-tb space-evenly"><div class="inner">horizontal-tb<br>space-evenly</div></div><div class="container horizontal-tb stretch"><div class="inner">horizontal-tb<br>stretch</div></div><div class="container horizontal-tb center"><div class="inner">horizontal-tb<br>center</div></div><div class="container horizontal-tb flex-start"><div class="inner">horizontal-tb<br>flex-start</div></div><div class="container horizontal-tb flex-end"><div class="inner">horizontal-tb<br>flex-end</div></div><div class="container horizontal-tb row"><div class="inner">horizontal-tb<br>row</div></div><div class="container horizontal-tb row-reverse"><div class="inner">horizontal-tb<br>row-reverse</div></div><div class="container horizontal-tb column"><div class="inner">horizontal-tb<br>column</div></div><div class="container vertical-rl column-reverse"><div class="inner">vertical-rl<br>column-reverse</div></div><div class="container vertical-rl start"><div class="inner">vertical-rl<br>start</div></div><div class="container vertical-rl center"><div class="inner">vertical-rl<br>center</div></div><div class="container vertical-rl space-around"><div class="inner">vertical-rl<br>space-around</div></div><div class="container vertical-rl space-between"><div class="inner">vertical-rl<br>space-between</div></div><div class="container vertical-rl space-evenly"><div class="inner">vertical-rl<br>space-evenly</div></div><div class="container vertical-rl stretch"><div class="inner">vertical-rl<br>stretch</div></div><div class="container vertical-rl center"><div class="inner">vertical-rl<br>center</div></div><div class="container vertical-rl flex-start"><div class="inner">vertical-rl<br>flex-start</div></div><div class="container vertical-rl flex-end"><div class="inner">vertical-rl<br>flex-end</div></div><div class="container vertical-rl row"><div class="inner">vertical-rl<br>row</div></div><div class="container vertical-rl row-reverse"><div class="inner">vertical-rl<br>row-reverse</div></div><div class="container vertical-rl column"><div class="inner">vertical-rl<br>column</div></div><div class="container vertical-lr column-reverse"><div class="inner">vertical-lr<br>column-reverse</div></div><div class="container vertical-lr start"><div class="inner">vertical-lr<br>start</div></div><div class="container vertical-lr center"><div class="inner">vertical-lr<br>center</div></div><div class="container vertical-lr space-around"><div class="inner">vertical-lr<br>space-around</div></div><div class="container vertical-lr space-between"><div class="inner">vertical-lr<br>space-between</div></div><div class="container vertical-lr space-evenly"><div class="inner">vertical-lr<br>space-evenly</div></div><div class="container vertical-lr stretch"><div class="inner">vertical-lr<br>stretch</div></div><div class="container vertical-lr center"><div class="inner">vertical-lr<br>center</div></div><div class="container vertical-lr flex-start"><div class="inner">vertical-lr<br>flex-start</div></div><div class="container vertical-lr flex-end"><div class="inner">vertical-lr<br>flex-end</div></div><div class="container vertical-lr row"><div class="inner">vertical-lr<br>row</div></div><div class="container vertical-lr row-reverse"><div class="inner">vertical-lr<br>row-reverse</div></div><div class="container vertical-lr column"><div class="inner">vertical-lr<br>column</div></div><div class="container sideways-rl column-reverse"><div class="inner">sideways-rl<br>column-reverse</div></div><div class="container sideways-rl start"><div class="inner">sideways-rl<br>start</div></div><div class="container sideways-rl center"><div class="inner">sideways-rl<br>center</div></div><div class="container sideways-rl space-around"><div class="inner">sideways-rl<br>space-around</div></div><div class="container sideways-rl space-between"><div class="inner">sideways-rl<br>space-between</div></div><div class="container sideways-rl space-evenly"><div class="inner">sideways-rl<br>space-evenly</div></div><div class="container sideways-rl stretch"><div class="inner">sideways-rl<br>stretch</div></div><div class="container sideways-rl center"><div class="inner">sideways-rl<br>center</div></div><div class="container sideways-rl flex-start"><div class="inner">sideways-rl<br>flex-start</div></div><div class="container sideways-rl flex-end"><div class="inner">sideways-rl<br>flex-end</div></div><div class="container sideways-rl row"><div class="inner">sideways-rl<br>row</div></div><div class="container sideways-rl row-reverse"><div class="inner">sideways-rl<br>row-reverse</div></div><div class="container sideways-rl column"><div class="inner">sideways-rl<br>column</div></div><div class="container sideways-lr column-reverse"><div class="inner">sideways-lr<br>column-reverse</div></div><div class="container sideways-lr start"><div class="inner">sideways-lr<br>start</div></div><div class="container sideways-lr center"><div class="inner">sideways-lr<br>center</div></div><div class="container sideways-lr space-around"><div class="inner">sideways-lr<br>space-around</div></div><div class="container sideways-lr space-between"><div class="inner">sideways-lr<br>space-between</div></div><div class="container sideways-lr space-evenly"><div class="inner">sideways-lr<br>space-evenly</div></div><div class="container sideways-lr stretch"><div class="inner">sideways-lr<br>stretch</div></div><div class="container sideways-lr center"><div class="inner">sideways-lr<br>center</div></div><div class="container sideways-lr flex-start"><div class="inner">sideways-lr<br>flex-start</div></div><div class="container sideways-lr flex-end"><div class="inner">sideways-lr<br>flex-end</div></div><div class="container sideways-lr row"><div class="inner">sideways-lr<br>row</div></div><div class="container sideways-lr row-reverse"><div class="inner">sideways-lr<br>row-reverse</div></div><div class="container sideways-lr column"><div class="inner">sideways-lr<br>column</div></div></div>

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    // Generating script.
    // const writingModes = [
    //   'horizontal-tb',
    //   'vertical-rl',
    //   'vertical-lr',
    //   'sideways-rl',
    //   'sideways-lr'
    // ];
    // const alignmentProps = [
    //   'start',
    //   'center',
    //   'space-around',
    //   'space-between',
    //   'space-evenly',
    //   'stretch',
    //   'center',
    //   'flex-start',
    //   'flex-end',
    //   'row',
    //   'row-reverse',
    //   'column',
    //   'column-reverse'
    // ];
    // for (let index in writingModes) {
    //   const writingMode = writingModes[index];
    //   const thisWriteMode = document.querySelector('#wrap').appendChild(makeContainer(writingMode));
    //   for (let index2 in alignmentProps) {
    //     const alignmentProp = alignmentProps[index2];
    //     // warning, type massage!
    //     if (index2 != (alignmentProps.length - 1)) {
    //       // Clone the first node of writing mode
    //       const dup = thisWriteMode.cloneNode()
    //       dup.className = dup.className + ' ' + alignmentProp;
    //       dup.appendChild(makeInner()).innerHTML = writingMode + '<br/>' + alignmentProp
    //       document.querySelector('#wrap').appendChild(dup)
    //     }
    //     else {
    //       // If this is the last alignmentProp we'll use the original container.
    //       thisWriteMode.className = thisWriteMode.className + ' ' + alignmentProp;
    //       thisWriteMode.appendChild(makeInner()).innerHTML = writingMode + '<br/>' + alignmentProp
    //     }
    //
    //   }
    //
    // }
    // function makeInner() {
    //   const inner = document.createElement('div')
    //   inner.className = 'inner'
    //   return inner
    // }
    //
    // function makeContainer(writingMode) {
    //   const container = document.createElement('div')
    //   container.className = `container ${writingMode}`;
    //   return container;
    // }

  </script>
  <script>
    test(function() {
      const container = document.querySelector('.horizontal-tb.start');
      const itemBR = container.querySelector('.inner').getBoundingClientRect();
      const containerBR = container.getBoundingClientRect();
      assert_equals(containerBR.top, itemBR.top);
      assert_equals(containerBR.left, itemBR.left);
    }, "Horizontal Top to Bottom Justified to Start.")

  </script>
</body>
</html>
